requirements.txt
Pillow
Pytesseract
tesseract